db: fix provides handling
Ensure that we only consider the provides when calculating the build
order and not otherwise.

Do so by have a separate providers table where we track the providers,
and add a separate `each_provider_for` function call so we can separate
out when we want include the providers and when we don't.

Also fixes a bug when the APKBUILD have more than one provider listed.

fixes commit 9e31086fa8f1 (db: include provides when calculate build
order)

fixes: https://gitlab.alpinelinux.org/alpine/lua-aports/-/issues/1